.
Basic gist of the app:
It's a study aid app. There is a page with Vocabulary Terms. As you click on the words, their definiton should come up on the same screen in a fragment. I am also implementing spinners to include a drop down menu on each word. These include options of how well the user understands the word.